feat: project scoped server option
