+++ packages/mcp-server-supabase/src/server.test.ts
@@ -34,6 +34,7 @@ beforeEach(async () => {
 
 type SetupOptions = {
   accessToken?: string;
+  projectId?: string;
   readOnly?: boolean;
 };
 
@@ -41,7 +42,7 @@ type SetupOptions = {
  * Sets up an MCP client and server for testing.
  */
 async function setup(options: SetupOptions = {}) {
-  const { accessToken = ACCESS_TOKEN, readOnly } = options;
+  const { accessToken = ACCESS_TOKEN, projectId, readOnly } = options;
   const clientTransport = new StreamTransport();
   const serverTransport = new StreamTransport();
 
@@ -63,6 +64,7 @@ async function setup(options: SetupOptions = {}) {
       apiUrl: API_URL,
       accessToken,
     },
+    projectId,
     readOnly,
   });
 
@@ -1909,5 +1911,117 @@ describe('tools', () => {
         );
       }
     }
+  });
+});
+
+describe('project scoped tools', () => {
+  test('no tool should accept a project_id', async () => {
+    const org = await createOrganization({
+      name: 'My Org',
+      plan: 'free',
+      allowed_release_channels: ['ga'],
+    });
+
+    const project = await createProject({
+      name: 'Project 1',
+      region: 'us-east-1',
+      organization_id: org.id,
+    });
+
+    const { client } = await setup({ projectId: project.id });
+
+    const result = await client.listTools();
+
+    expect(result.tools).toBeDefined();
+    expect(Array.isArray(result.tools)).toBe(true);
+
+    for (const tool of result.tools) {
+      const schemaProperties = tool.inputSchema.properties ?? {};
+      expect(
+        'project_id' in schemaProperties,
+        `tool ${tool.name} should not accept a project_id`
+      ).toBe(false);
+    }
+  });
+
+  test('invalid project ID should throw an error', async () => {
+    const { callTool } = await setup({ projectId: 'invalid-project-id' });
+
+    const listTablesPromise = callTool({
+      name: 'list_tables',
+      arguments: {
+        schemas: ['public'],
+      },
+    });
+
+    await expect(listTablesPromise).rejects.toThrow('Project not found');
+  });
+
+  test('passing project_id to a tool should throw an error', async () => {
+    const org = await createOrganization({
+      name: 'My Org',
+      plan: 'free',
+      allowed_release_channels: ['ga'],
+    });
+
+    const project = await createProject({
+      name: 'Project 1',
+      region: 'us-east-1',
+      organization_id: org.id,
+    });
+    project.status = 'ACTIVE_HEALTHY';
+
+    const { callTool } = await setup({ projectId: project.id });
+
+    const listTablesPromise = callTool({
+      name: 'list_tables',
+      arguments: {
+        project_id: 'my-project-id',
+        schemas: ['public'],
+      },
+    });
+
+    await expect(listTablesPromise).rejects.toThrow('Unrecognized key');
+  });
+
+  test('listing tables implicitly uses the scoped project_id', async () => {
+    const org = await createOrganization({
+      name: 'My Org',
+      plan: 'free',
+      allowed_release_channels: ['ga'],
+    });
+
+    const project = await createProject({
+      name: 'Project 1',
+      region: 'us-east-1',
+      organization_id: org.id,
+    });
+    project.status = 'ACTIVE_HEALTHY';
+
+    project.db
+      .sql`create table test (id integer generated always as identity primary key)`;
+
+    const { callTool } = await setup({ projectId: project.id });
+
+    const result = await callTool({
+      name: 'list_tables',
+      arguments: {
+        schemas: ['public'],
+      },
+    });
+
+    expect(result).toEqual([
+      expect.objectContaining({
+        name: 'test',
+        schema: 'public',
+        columns: [
+          expect.objectContaining({
+            name: 'id',
+            is_identity: true,
+            is_generated: false,
+          }),
+        ],
+      }),
+    ]);
   });
 });
